requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQL==1.5.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
PyMySQL==1.1.0
Werkzeug==3.0.0
python-decouple==3.8
